rgAddrMap:  Mapped Address cache key is 32-bit documentation address.
    Prior key was the 64-bit real IO address resulting in compiler warnings
    on  std::map<uint64_t,void*>.  No behavior change.
    See:  doc/address_map.text

+++ librgpio/src/rgAddrMap.h
@@ -20,7 +20,7 @@ class rgAddrMap {
     const char		*ModeStr;	// memory mode string, NULL= unset,
 					// "/dev/mem", "/dev/gpiomem", "fake_mem"
 
-    std::map<uint64_t,void*>  BlkCache;	// cache of mapped memory blocks
+    std::map<uint32_t,void*>  BlkCache;	// cache of mapped memory blocks
 
 //  int			Prot;		// mmap() prot field
     bool		Debug;		// debug trace on stderr
